refactor: move lineparse into specparse
