+++ src/plccng/specparse/lexparse/lexvalidate/check_format_of_names.py
@@ -1,8 +1,7 @@
 import re
 from dataclasses import dataclass
 
-from plccng.specparse.ValidationError import ValidationError
-
+from ...ValidationError import ValidationError
 from ..LexicalRule import LexicalRule
