+++ src/plccng/specparse/lineparse/fromstring.py
@@ -1,4 +1,4 @@
-from plccng.lineparse.fromstrings import fromstrings
+from ..lineparse.fromstrings import fromstrings
 
 
 def fromstring(string, file=None, startLineNumber=1):
