+++ src/plccng/specparse/roughparse/parse_blocks_test.py
@@ -1,8 +1,7 @@
 from pytest import raises
 
-import plccng.lineparse as lineparse
+from .. import lineparse
 from .Block import Block
-
 from .parse_blocks import parse_blocks
 from .UnclosedBlockError import UnclosedBlockError
