+++ src/plccng/specparse/roughparse/parse_dividers.py
@@ -1,7 +1,6 @@
 import re
 
-from plccng.lineparse.Line import Line
-
+from ..lineparse.Line import Line
 from .Divider import Divider
