+++ src/plccng/specparse/roughparse/parse_includes_test.py
@@ -1,7 +1,6 @@
-import plccng.lineparse as lineparse
+from .. import lineparse
 from .Block import Block
 from .Include import Include
-
 from .parse_blocks import parse_blocks
 from .parse_includes import parse_includes
