+++ src/plccng/specparse/roughparse/resolve_includes.py
@@ -1,7 +1,6 @@
 from pathlib import Path
 
-import plccng.lineparse as lineparse
-
+from .. import lineparse
 from .CircularIncludeError import CircularIncludeError
 from .Include import Include
 from .parse_blocks import parse_blocks
