+++ src/plccng/specparse/semparse/UndefinedTargetLocatorError.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from plccng.specparse.ValidationError import ValidationError
+from ..ValidationError import ValidationError
 
 
 @dataclass
