+++ src/plccng/specparse/semparse/parse_target_locator_test.py
@@ -1,7 +1,6 @@
 from pytest import raises
 
-from plccng.lineparse import Line
-
+from ..lineparse import Line
 from .parse_target_locator import InvalidTargetLocatorError, parse_target_locator
 from .TargetLocator import TargetLocator
